+++ com.unity.render-pipelines.lightweight/Runtime/IRendererSetup.cs
@@ -2,8 +2,8 @@ using UnityEngine.Rendering.LWRP;
 
 namespace UnityEngine.Experimental.Rendering.LWRP
 {
-    public interface IRendererSetup
+    public abstract class IRendererSetup
     {
-        void Setup(ScriptableRenderer renderer, ref RenderingData renderingData);
+        public abstract void Setup(ScriptableRenderer renderer, ref RenderingData renderingData);
     }
 }