+++ com.unity.testing.srp.lightweight/Tests/Scenes/052_LWCallbacks/CameraCallbackTests.cs
@@ -20,8 +20,14 @@ public class CameraCallbackTests : MonoBehaviour
 	static RenderTargetHandle afterTransparent;
 	static RenderTargetHandle afterAll;
 
+    Material m_CopyDepthMaterial;
+    Material m_SamplingMaterial;
+    
 	public CameraCallbackTests()
 	{
+	    m_CopyDepthMaterial = CoreUtils.CreateEngineMaterial(Shader.Find("Hidden/Lightweight Render Pipeline/CopyDepth"));
+	    m_SamplingMaterial = CoreUtils.CreateEngineMaterial(Shader.Find("Hidden/Lightweight Render Pipeline/Sampling"));
+	    
 		beforeAll.Init("_BeforeAll");
 		afterOpaque.Init("_AfterOpaque");
 		afterOpaquePost.Init("_AfterOpaquePost");
@@ -63,7 +69,7 @@ public class CameraCallbackTests : MonoBehaviour
 	ScriptableRenderPass IAfterOpaquePass.GetPassToEnqueue(RenderTextureDescriptor baseDescriptor, RenderTargetHandle colorAttachmentHandle,
 		RenderTargetHandle depthAttachmentHandle)
 	{
-		var pass = new CopyColorPass();
+		var pass = new CopyColorPass(m_SamplingMaterial);
 		pass.Setup(colorAttachmentHandle, afterOpaque);
 		return pass;
 	}
@@ -71,7 +77,7 @@ public class CameraCallbackTests : MonoBehaviour
 	ScriptableRenderPass IAfterOpaquePostProcess.GetPassToEnqueue(RenderTextureDescriptor baseDescriptor, RenderTargetHandle colorHandle,
 		RenderTargetHandle depthHandle)
 	{
-		var pass = new CopyColorPass();;
+		var pass = new CopyColorPass(m_SamplingMaterial);;
 		pass.Setup(colorHandle, afterOpaquePost);
 		return pass;
 	}
@@ -79,7 +85,7 @@ public class CameraCallbackTests : MonoBehaviour
 	ScriptableRenderPass IAfterSkyboxPass.GetPassToEnqueue(RenderTextureDescriptor baseDescriptor, RenderTargetHandle colorHandle,
 		RenderTargetHandle depthHandle)
 	{
-		var pass = new CopyColorPass();
+		var pass = new CopyColorPass(m_SamplingMaterial);
 		pass.Setup(colorHandle, afterSkybox);
 		return pass;
 	}
@@ -87,7 +93,7 @@ public class CameraCallbackTests : MonoBehaviour
 	ScriptableRenderPass IAfterTransparentPass.GetPassToEnqueue(RenderTextureDescriptor baseDescriptor, RenderTargetHandle colorHandle,
 		RenderTargetHandle depthHandle)
 	{
-		var pass = new CopyColorPass();
+		var pass = new CopyColorPass(m_SamplingMaterial);
 		pass.Setup(colorHandle, afterTransparent);
 		return pass;
 	}
@@ -96,11 +102,13 @@ public class CameraCallbackTests : MonoBehaviour
 	{
         private RenderTargetHandle colorHandle;
         private RenderTargetHandle depthHandle;
+	    Material m_BlitMaterial;
 
         public BlitPass(RenderTargetHandle colorHandle, RenderTargetHandle depthHandle)
         {
             this.colorHandle = colorHandle;
             this.depthHandle = colorHandle;
+            m_BlitMaterial = CoreUtils.CreateEngineMaterial(Shader.Find("Hidden/Lightweight Render Pipeline/Blit"));
         }
 
         public override void Execute(ScriptableRenderer renderer, ScriptableRenderContext context, ref RenderingData renderingData)
@@ -108,11 +116,9 @@ public class CameraCallbackTests : MonoBehaviour
 			if (renderer == null)
 				throw new ArgumentNullException("renderer");
 
-		    var pass = new CopyColorPass();
+		    var pass = new CopyColorPass(CoreUtils.CreateEngineMaterial(Shader.Find("Hidden/Lightweight Render Pipeline/Sampling")));
 		    pass.Setup(colorHandle, afterAll);
             pass.Execute(renderer, context, ref renderingData);
-
-            Material material = renderer.GetMaterial(MaterialHandle.Blit);
 
 			CommandBuffer cmd = CommandBufferPool.Get("Blit Pass");
 			cmd.SetRenderTarget(colorHandle.id, depthHandle.id);
@@ -120,27 +126,27 @@ public class CameraCallbackTests : MonoBehaviour
 			
 			cmd.SetViewport(new Rect(0, renderingData.cameraData.camera.pixelRect.height / 2.0f, renderingData.cameraData.camera.pixelRect.width / 3.0f, renderingData.cameraData.camera.pixelRect.height / 2.0f));
 			cmd.SetGlobalTexture("_BlitTex", beforeAll.Identifier());
-		    ScriptableRenderer.RenderFullscreenQuad(cmd, material);
+		    ScriptableRenderer.RenderFullscreenQuad(cmd, m_BlitMaterial);
 			
 			cmd.SetViewport(new Rect(renderingData.cameraData.camera.pixelRect.width / 3.0f, renderingData.cameraData.camera.pixelRect.height / 2.0f, renderingData.cameraData.camera.pixelRect.width / 3.0f, renderingData.cameraData.camera.pixelRect.height / 2.0f));
 			cmd.SetGlobalTexture("_BlitTex", afterOpaque.Identifier());
-		    ScriptableRenderer.RenderFullscreenQuad(cmd, material);
+		    ScriptableRenderer.RenderFullscreenQuad(cmd, m_BlitMaterial);
 			
 			cmd.SetViewport(new Rect(renderingData.cameraData.camera.pixelRect.width / 3.0f * 2.0f, renderingData.cameraData.camera.pixelRect.height / 2.0f, renderingData.cameraData.camera.pixelRect.width / 3.0f, renderingData.cameraData.camera.pixelRect.height / 2.0f));
 			cmd.SetGlobalTexture("_BlitTex", afterOpaquePost.Identifier());
-		    ScriptableRenderer.RenderFullscreenQuad(cmd, material);			
+		    ScriptableRenderer.RenderFullscreenQuad(cmd, m_BlitMaterial);			
 						
 			cmd.SetViewport(new Rect(0f, 0f, renderingData.cameraData.camera.pixelRect.width / 3.0f, renderingData.cameraData.camera.pixelRect.height / 2.0f));
 			cmd.SetGlobalTexture("_BlitTex", afterSkybox.Identifier());
-		    ScriptableRenderer.RenderFullscreenQuad(cmd, material);
+		    ScriptableRenderer.RenderFullscreenQuad(cmd, m_BlitMaterial);
 			
 			cmd.SetViewport(new Rect(renderingData.cameraData.camera.pixelRect.width / 3.0f, 0f, renderingData.cameraData.camera.pixelRect.width / 3.0f, renderingData.cameraData.camera.pixelRect.height / 2.0f));
 			cmd.SetGlobalTexture("_BlitTex", afterTransparent.Identifier());
-		    ScriptableRenderer.RenderFullscreenQuad(cmd, material);
+		    ScriptableRenderer.RenderFullscreenQuad(cmd, m_BlitMaterial);
 			
 			cmd.SetViewport(new Rect(renderingData.cameraData.camera.pixelRect.width / 3.0f * 2.0f, 0f, renderingData.cameraData.camera.pixelRect.width / 3.0f, renderingData.cameraData.camera.pixelRect.height / 2.0f));
 			cmd.SetGlobalTexture("_BlitTex", afterAll.Identifier());
-		    ScriptableRenderer.RenderFullscreenQuad(cmd, material);
+		    ScriptableRenderer.RenderFullscreenQuad(cmd, m_BlitMaterial);
 
             context.ExecuteCommandBuffer(cmd);
 			CommandBufferPool.Release(cmd);